[Refactor] Move Ts exporter to own package
