+++ src/exporter/typescript/argument.go
@@ -1,0 +1,9 @@
+package typescript
+
+import (
+	"github.com/maasasia/donggu/dictionary"
+)
+
+type ArgumentFormatter interface {
+	Format(key string, format dictionary.TemplateKeyFormat) string
+}
